A partial port of the libvpx vp8 encoder to java + webm improvements
- The MKV muxer is now now waiting to write the file to the disk till
the finish() call. Instead it writes after every keyframe. This allows
larger files
- Preexisting vpx code was merged with the new libvpx code
- A new yuv to webm transcoder sample was added
- The libvpx port allows multiple threads to encode in parralel with
multiple encoder objects (but a single one does not do multi-threading)
- Here are the features that were not ported:
-- CPU specific optimisations
-- VP9 encoder/decoder
-- VP8 decoder
-- VP8 encoder postprocessing, spatial-resampling, error-concealment,
multi-resolution-encoding, temporal-denoising and multi threading
- Many constructs for the above omitted features are still partially
ported, they are just not tested and only used to the extent that the
basic features of the encoder needed it.

+++ src/main/java/org/jcodec/api/transcode/SinkImpl.java
@@ -191,7 +191,7 @@ public class SinkImpl implements Sink, PacketSink {
             } else if (Codec.H264 == outputVideoCodec) {
                 videoEncoder = createH264Encoder();
             } else if (VP8 == outputVideoCodec) {
-                videoEncoder = VP8Encoder.createVP8Encoder(10);
+                videoEncoder = VP8Encoder.createVP8Encoder((short)10);
             } else if (PNG == outputVideoCodec) {
                 videoEncoder = new PNGEncoder();
             } else if (Codec.RAW == outputVideoCodec) {
